Now made public
A more complex version of my previous donut project. This rotating donut is written in python. This donut uses the same principle as my previous donut project but instead of covering a small portion of the screen this donut has a 16:9 ratio. A bonus feature is that it changes colour.